✨feat:add git ssh key check
