+++ packages/utils/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./ui";
 export * from "./types";
+export * from "./exec";
