+++ packages/web-tools/src/features/index.ts
@@ -2,4 +2,5 @@ export default [
   require.resolve("./checkNode"),
   require.resolve("./checkChrome"),
   require.resolve("./checkDefaultBrowser"),
+  require.resolve("./checkGit"),
 ];
